.requirements.txt
asgiref==3.7.2
Django==4.2.3
django-ckeditor==6.6.1
django-js-asset==2.1.0
djangorestframework==3.14.0
Faker==19.2.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
